Addressing review suggestions
